Filter and profile creation flow working.
You can now login through sms, create a profile, and set your initial filters. And it all works with the backend support.

+++ pull_common/lib/src/model/filters.dart
@@ -3,8 +3,8 @@ class Filters{
   bool menChecked = false;
   bool womenChecked = false;
   bool nonBinaryChecked = false;
-  double lowerAge = 18;
-  double upperAge = 100;
+  int lowerAge = 18;
+  int upperAge = 100;
   double lowerHeight = 55;
   double upperHeight = 275;
   double maxDistance = 15;
